# Conflicts:
#	BackEnd/service/src/main/java/com/basis/sge/service/servico/EmailServico.java
#	BackEnd/service/src/main/resources/config/application.yml

+++ BackEnd/service/src/main/java/com/basis/sge/service/servico/EventoServico.java
@@ -5,15 +5,17 @@ import com.basis.sge.service.dominio.Evento;
 import com.basis.sge.service.repositorio.EventoRepositorio;
 import com.basis.sge.service.servico.dto.EventoDTO;
 
+import com.basis.sge.service.servico.dto.TipoEventoDTO;
 import com.basis.sge.service.servico.exception.RegraNegocioException;
 import com.basis.sge.service.servico.mapper.EventoMapper;
 import lombok.RequiredArgsConstructor;
 
-import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+//import org.springframework.transaction.annotation.Transactional;
 
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
@@ -34,12 +36,15 @@ public class EventoServico {
 
     public EventoDTO obterPorId(Integer id){
         Evento evento = eventoRepositorio.findById(id)
-                .orElseThrow(() -> new RegraNegocioException("Usuario não Existe"));
+                .orElseThrow(() -> new RegraNegocioException("Evento não Existe"));
         return eventoMapper.toDto(evento);
     }
 
 
     public EventoDTO criar(EventoDTO eventoDTO) {
+        if (eventoRepositorio.existsByTitulo(eventoDTO.getTitulo())){
+            throw new RegraNegocioException("Um evento com esse titulo já existe");
+        }
         Evento evento = eventoMapper.toEntity(eventoDTO);
         Evento eventoSalvo = eventoRepositorio.save(evento);
 
@@ -47,7 +52,9 @@ public class EventoServico {
     }
 
     public EventoDTO atualizar(EventoDTO eventoDTO) {
-
+        if (eventoRepositorio.existsByTituloAndIdNot(eventoDTO.getTitulo(),eventoDTO.getId())){
+            throw new RegraNegocioException("Um evento com esse titulo já existe");
+        }
         Evento evento = eventoMapper.toEntity(eventoDTO);
         Evento eventoAtualizado = eventoRepositorio.save(evento);
 
@@ -56,6 +63,9 @@ public class EventoServico {
 
 
     public void deletar(Integer id) {
+        if(!eventoRepositorio.existsById(id)){
+            throw new RegraNegocioException("Evento Não Existe");
+        }
         eventoRepositorio.deleteById(id);
     }
